+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 Unreleased
 ==========
 
+- Fix ignored formatting while using ``logging.Handler`` sinks (`#4 <https://github.com/Delgan/loguru/issues/4>`_)
 - Fix impossibility to set empty environment variable color on Windows (`#7 <https://github.com/Delgan/loguru/issues/7>`_)